Fix TranslatableAdmin get_object() to not use fallbacks.
The fix for [Nani issue 41](https://github.com/ojii/django-nani/issues/41)
results in TranslatableAdmin returning an object with a different translation,
instead of no translation, when editing an object in the admin. Applying
changes made on a form can then apply them to the wrong translation.

If this change is omitted, then the real fix for TranslatableModelForm
breaks the admin and this is caught by the tests.