Fix #141: TranslatableModelAdmin uses absolute URLs for language tabs, breaks reverse proxies
On the language tabs of our admin, behind a reverse proxy, we see this:

    <input type="hidden" class="language_button selected" name="en" />
    <span class="current">English</span>
    <span class="empty"><a href="http://fen-vz-inaspauthoraid.fen.aptivate.org/en/admin/news/newsitem/add/?language=es">Spanish</a> </span>

`fen-vz-inaspauthoraid.fen.aptivate.org` is an internal hostname which cannot
be accessed without a VPN. It's not the URL that we used to load the page,
through the reverse proxy.

It's not necessary to include a host and protocol in the rendered links. You
can save a little bandwidth (page size) and avoid this problem by using
relative links:

    url = '%s?%s' % (request.path, urlencode(get))